Adding a ton of assets
Changing sprites to Secret of Mana 2 over Legend of Mana. They are an
order of magnitude easier to work with. I changed the data to reflect
these, I made it so that players and monsters are anchored in the center
because the sprite sheets do not have both left and right sprites. The
consequence of this was that I also had to change the camera to no
longer reflect sprite width and height (which is good). I also had to
change all instances of collision detection to use the player bounds
(which account for the anchor) instead of just the player. I also added
Z and X to input and made it so that you can run and walk.